split handlers into two files for client and server
